requirements.txt
alembic==1.0.5
asn1crypto==0.24.0
cffi==1.11.5
Click==7.0
cryptography==2.4.2
Flask==1.0.2
Flask-Cors==3.0.3
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
paginate==0.5.6
pycparser==2.19
PyMySQL==0.9.2
python-dateutil==2.7.5
python-editor==1.0.3
requests==2.21.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
six==1.12.0
SQLAlchemy==1.2.15
Werkzeug==0.14.1
PyJWT==1.4.2
